test(editor-assist): split route tests by responsibility
